Rename `step_converter.rs` to `angular_step_converter.rs`
