+++ control-core/src/converters/mod.rs
@@ -1,2 +1,2 @@
 pub mod linear_step_converter;
-pub mod step_converter;
+pub mod angular_step_converter;
